+++ AmazonChimeSDKDemo/AmazonChimeSDKDemo/BroadcastScreenCaptureModel.swift
@@ -13,6 +13,7 @@ class BroadcastScreenCaptureModel {
     let meetingSessionConfig: MeetingSessionConfiguration
     let appGroupUserDefaults = UserDefaults(suiteName: AppConfiguration.appGroupId)
     let userDefaultsKeyMeetingId = "demoMeetingId"
+    let userDefaultsKeyExternalMeetingId = "demoExternalMeetingId"
     let userDefaultsKeyCredentials = "demoMeetingCredentials"
     let userDefaultsKeyUrls = "demoMeetingUrls"
     let logger = ConsoleLogger(name: "BroadcastScreenCaptureModel")
@@ -44,6 +45,7 @@ class BroadcastScreenCaptureModel {
             return
         }
         appGroupUserDefaults.set(meetingSessionConfig.meetingId, forKey: userDefaultsKeyMeetingId)
+        appGroupUserDefaults.set(meetingSessionConfig.externalMeetingId, forKey: userDefaultsKeyExternalMeetingId)
         let encoder = JSONEncoder()
         if let credentials = try? encoder.encode(meetingSessionConfig.credentials) {
             appGroupUserDefaults.set(credentials, forKey: userDefaultsKeyCredentials)
@@ -59,6 +61,7 @@ class BroadcastScreenCaptureModel {
             return
         }
         appGroupUserDefaults.removeObject(forKey: userDefaultsKeyMeetingId)
+        appGroupUserDefaults.removeObject(forKey: userDefaultsKeyExternalMeetingId)
         appGroupUserDefaults.removeObject(forKey: userDefaultsKeyCredentials)
         appGroupUserDefaults.removeObject(forKey: userDefaultsKeyUrls)
     }