+++ CHANGELOG.md
@@ -2,9 +2,18 @@
 
 ### Added
 * Added support for XCFramework. AmazonChimeSDK and AmazonChimeSDKMedia binaries now contain .xcframework format as well. See `README.md` for updated Setup instructions.
+* Added Analytics
+    * `EventAnalyticsController`, `EventAnalyticsFacade`, `EventAnalyticsObserver` to handle analytics.
+    * Added `EventAttributesName`, `EventName` for meeting event information.
+    * **Breaking** Added `EventAnalyticsController` to parameters of `DefaultAudioVideoFacade`.
+    * Added `EventAnalyticsController` to parameters of `DefaultCameraCaptureSource`, `DefaultAudioClientObserver`, `DefaultAudioClientController`.
+    * Added `externalMeetingId` to property of `MeetingSessionConfiguration`.
+    * Added `EventAnalyticsController` to property of `MeetingSession`.
+    * [Demo] Added `PostLogger` to demonstrate making HTTP POST request to server with logs.
+    * [Documentation] Added usage guide for analytics
 
 ## Fixed
-* Fixed data message conversion that sometimes does not handle null terminator when converting from string to c-string.
+* Fixed data message conversion that sometimes does not handle null terminator when converting from string to c-string. (Issue #217)
 
 ## [0.14.0] - 2021-01-21